# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AUZEF Chat Analiz API",
    "description": "AUZEF chatbot mesajlarından sık sorulan soru ve tema analizi.\n\nTüm hata cevapları RFC 9457 Problem Details biçimindedir ve her zaman `type`, `title`, `status`, `code`, `detail`, `trace_id` alanlarını taşır. `retry_after` YALNIZCA 429 cevaplarında bulunur ve başka hiçbir cevapta `null` olarak dahi yer almaz.\n\nTüm tarihler UTC ISO 8601'dir ve `Z` ile biter (`YYYY-MM-DDTHH:MM:SS.sssZ`).",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/analyses": {
      "post": {
        "tags": [
          "analyses"
        ],
        "summary": "Analizi başlat ve job'ı kuyruğa al",
        "operationId": "create_analysis_api_v1_analyses_post",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "description": "Opsiyonel. Aynı anahtar + aynı gövde ile tekrarlanan istek, ilk isteğin 202 cevabını aynen döndürür (replay). Aynı anahtar + FARKLI gövde ise 409 `JOB_CONFLICT` üretir. Kayıt 24 saat saklanır.\n\nReplay, orijinal 202'yi döndürür ve istekle gelen yeni `X-OpenRouter-Key` header'ını YOK SAYAR: anahtar orijinal job'a bağlıdır. Aksi hâlde anahtar rotasyonundan sonraki bir retry, ölmüş bir job'ı sessizce diriltirdi.\n\nFingerprint ve saklama anahtarı ADR-0002 #3'te byte seviyesinde tanımlıdır; secret/header değerleri fingerprint'e girmez.",
            "schema": {
              "type": "string",
              "maxLength": 255
            },
            "example": "8f14e45f-ceea-467a-9f6b-2c1d3e4a5b6c"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalysisRequest"
              },
              "examples": {
                "Tipik analiz isteği": {
                  "value": {
                    "upload_id": "3f2504e0-4f89-41d3-9a0c-0305e82c3301",
                    "sheet_name": "Mesajlar",
                    "text_column": "mesaj",
                    "model": "google/gemini-2.5-flash",
                    "prompt_version": "faq_analysis/v1",
                    "top_n": 8,
                    "max_cost_usd": 10.0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisCreated"
                },
                "examples": {
                  "Kuyruğa alındı": {
                    "value": {
                      "analysis_id": "6b1cf3d2-0a44-4f1b-9d64-1c2a7e5f8b90",
                      "status": "queued"
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "404": {
            "description": "RFC 9457 Problem Details — olası kodlar: JOB_NOT_FOUND",
            "x-error-codes": [
              "JOB_NOT_FOUND"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "JOB_NOT_FOUND": {
                    "value": {
                      "type": "/errors/job-not-found",
                      "title": "İşlem bulunamadı",
                      "status": 404,
                      "code": "JOB_NOT_FOUND",
                      "detail": "İşlem bulunamadı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "409": {
            "description": "RFC 9457 Problem Details — olası kodlar: JOB_CONFLICT",
            "x-error-codes": [
              "JOB_CONFLICT"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "JOB_CONFLICT": {
                    "value": {
                      "type": "/errors/job-conflict",
                      "title": "İşlem durumu bu isteğe uygun değil",
                      "status": 409,
                      "code": "JOB_CONFLICT",
                      "detail": "İşlem durumu bu isteğe uygun değil (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "422": {
            "description": "RFC 9457 Problem Details — olası kodlar: COST_LIMIT_EXCEEDED, INVALID_MODEL, INVALID_PROMPT, PROVIDER_AUTH_FAILED, REQUEST_VALIDATION, SHEET_OR_COLUMN_NOT_FOUND",
            "x-error-codes": [
              "REQUEST_VALIDATION",
              "INVALID_MODEL",
              "INVALID_PROMPT",
              "COST_LIMIT_EXCEEDED",
              "PROVIDER_AUTH_FAILED",
              "SHEET_OR_COLUMN_NOT_FOUND"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "REQUEST_VALIDATION": {
                    "value": {
                      "type": "/errors/request-validation",
                      "title": "İstek doğrulanamadı",
                      "status": 422,
                      "code": "REQUEST_VALIDATION",
                      "detail": "İstek doğrulanamadı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "INVALID_MODEL": {
                    "value": {
                      "type": "/errors/invalid-model",
                      "title": "Geçersiz model",
                      "status": 422,
                      "code": "INVALID_MODEL",
                      "detail": "Geçersiz model (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "INVALID_PROMPT": {
                    "value": {
                      "type": "/errors/invalid-prompt",
                      "title": "Geçersiz prompt sürümü",
                      "status": 422,
                      "code": "INVALID_PROMPT",
                      "detail": "Geçersiz prompt sürümü (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "COST_LIMIT_EXCEEDED": {
                    "value": {
                      "type": "/errors/cost-limit-exceeded",
                      "title": "Maliyet sınırı aşıldı",
                      "status": 422,
                      "code": "COST_LIMIT_EXCEEDED",
                      "detail": "Maliyet sınırı aşıldı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "PROVIDER_AUTH_FAILED": {
                    "value": {
                      "type": "/errors/provider-auth-failed",
                      "title": "Sağlayıcı kimlik doğrulaması başarısız",
                      "status": 422,
                      "code": "PROVIDER_AUTH_FAILED",
                      "detail": "Sağlayıcı kimlik doğrulaması başarısız (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "SHEET_OR_COLUMN_NOT_FOUND": {
                    "value": {
                      "type": "/errors/sheet-or-column-not-found",
                      "title": "Sayfa veya kolon bulunamadı",
                      "status": 422,
                      "code": "SHEET_OR_COLUMN_NOT_FOUND",
                      "detail": "Sayfa veya kolon bulunamadı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  },
                  "REQUEST_VALIDATION (alan yolu olmayan hata)": {
                    "value": {
                      "type": "/errors/request-validation",
                      "title": "İstek doğrulanamadı",
                      "status": 422,
                      "code": "REQUEST_VALIDATION",
                      "detail": "İstek gövdesi veya parametreleri doğrulanamadı.",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": [
                        {
                          "field": null,
                          "message": "Field required"
                        },
                        {
                          "field": "top_n",
                          "message": "Input should be greater than or equal to 1"
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "500": {
            "description": "RFC 9457 Problem Details — olası kodlar: INTERNAL_ERROR",
            "x-error-codes": [
              "INTERNAL_ERROR"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "INTERNAL_ERROR": {
                    "value": {
                      "type": "/errors/internal-error",
                      "title": "Beklenmeyen hata",
                      "status": 500,
                      "code": "INTERNAL_ERROR",
                      "detail": "Beklenmeyen hata (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          },
          "501": {
            "description": "RFC 9457 Problem Details — olası kodlar: NOT_IMPLEMENTED",
            "x-error-codes": [
              "NOT_IMPLEMENTED"
            ],
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                },
                "examples": {
                  "NOT_IMPLEMENTED": {
                    "value": {
                      "type": "/errors/not-implemented",
                      "title": "Uç nokta henüz uygulanmadı",
                      "status": 501,
                      "code": "NOT_IMPLEMENTED",
                      "detail": "Uç nokta henüz uygulanmadı (örnek gövde).",
                      "trace_id": "9d8c7b6a-5e4f-4321-8abc-0123456789ab",
                      "errors": []
                    }
                  }
                }
              }
            },
            "headers": {
              "X-Trace-Id": {
                "description": "Bu isteğin izleme kimliği; hata gövdesindeki trace_id ile aynı.",
                "schema": {
                  "type": "string",
                  "format": "uuid"
                },
                "example": "9d8c7b6a-5e4f-4321-8abc-0123456789ab"
              }
            }
          }
        },
        "security": [
          {
            "OpenRouterKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AnalysisCreated": {
        "properties": {
          "analysis_id": {
            "type": "string",
            "format": "uuid",
            "title": "Analysis Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status",
            "default": "queued"
          }
        },
        "type": "object",
        "required": [
          "analysis_id",
          "status"
        ],
        "title": "AnalysisCreated",
        "description": "POST /api/v1/analyses — 202 Accepted."
      },
...
      "AnalysisRequest": {
        "properties": {
          "upload_id": {
            "type": "string",
            "format": "uuid",
            "title": "Upload Id"
          },
          "sheet_name": {
            "type": "string",
            "minLength": 1,
            "title": "Sheet Name"
          },
          "text_column": {
            "type": "string",
            "minLength": 1,
            "title": "Text Column"
          },
          "model": {
            "$ref": "#/components/schemas/ModelId"
          },
          "prompt_version": {
            "$ref": "#/components/schemas/PromptVersion"
          },
          "top_n": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 1.0,
            "title": "Top N"
          },
          "max_cost_usd": {
            "type": "number",
            "maximum": 100.0,
            "exclusiveMinimum": 0.0,
            "title": "Max Cost Usd"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "upload_id",
          "sheet_name",
          "text_column",
          "model",
          "prompt_version",
          "top_n",
          "max_cost_usd"
        ],
        "title": "AnalysisRequest",
        "description": "POST /api/v1/analyses gövdesi.\n\nOpenRouter anahtarı BU GÖVDEDE YER ALMAZ — yalnızca `X-OpenRouter-Key`\nheader'ında taşınır (ADR-0001 §6/§9). `RequestValidationError` handler'ı\n`errors[].input` alanını asla yankılamaz; anahtar yanlışlıkla gövdeye\nkonursa geri sızdırılmaması için (ADR-0002 #7)."
      },
...
      "ErrorCode": {
        "type": "string",
        "enum": [
          "UPLOAD_TOO_LARGE",
          "UPLOAD_INVALID_TYPE",
          "UPLOAD_CORRUPT_OR_ENCRYPTED",
          "SHEET_OR_COLUMN_NOT_FOUND",
          "REQUEST_VALIDATION",
          "INVALID_MODEL",
          "INVALID_PROMPT",
          "COST_LIMIT_EXCEEDED",
          "PROVIDER_AUTH_FAILED",
          "PROVIDER_RATE_LIMITED",
          "PROVIDER_BAD_RESPONSE",
          "PROVIDER_TIMEOUT",
          "JOB_NOT_FOUND",
          "JOB_CONFLICT",
          "NOT_IMPLEMENTED",
          "SERVICE_NOT_READY",
          "INTERNAL_ERROR"
        ],
        "title": "ErrorCode",
        "description": "Dondurulmuş hata sözlüğü ve BE-02 readiness genişletmesi."
      },
...
      "ModelId": {
        "type": "string",
        "enum": [
          "anthropic/claude-sonnet-4.6",
          "openai/gpt-4.1-mini",
          "google/gemini-2.5-flash"
        ],
        "title": "ModelId",
        "description": "BE-01'de dondurulmuş, structured-output destekli model whitelist'i."
      },
...
      "ProblemDetails": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "integer",
            "title": "Status"
          },
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          },
          "trace_id": {
            "type": "string",
            "format": "uuid",
            "title": "Trace Id"
          },
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors"
          },
          "retry_after": {
            "type": "number",
            "minimum": 0.0,
            "title": "Retry After"
          }
        },
        "type": "object",
        "required": [
          "type",
          "title",
          "status",
          "code",
          "detail",
          "trace_id",
          "errors"
        ],
        "title": "ProblemDetails",
        "description": "ADR-0001 §7 / ADR-0002 #6.\n\n`type`, `title`, `status`, `code`, `detail` ve `trace_id` HER cevapta\nbulunur — frontend'in `problemDetailsSchema`'sı altısını da zorunlu tutar.\n\nBu garantiyi veren `json_schema_serialization_defaults_required` artık\n`ApiModel`'de; burada tekrarlanmıyor. Bayrağın YALNIZCA bu modelde olması,\ndiğer sekiz cevap modelinin openapi.json'da yanlış `required` listesiyle\nyayımlanmasının sebebiydi."
      },
      "PromptVersion": {
        "type": "string",
        "enum": [
          "faq_analysis/v1"
        ],
        "title": "PromptVersion",
        "description": "Backend'de sürümlenmiş ve dağıtıma dâhil prompt'lar."
      },
...
    },
    "securitySchemes": {
      "OpenRouterKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-OpenRouter-Key",
        "description": "BYOK OpenRouter anahtarı. Yalnızca bu header'da taşınır; gövdeye veya sorgu parametresine konmaz, loglarda redakte edilir ve kalıcı olarak saklanmaz."
      }
    }
  }
}